khr_get_memory_requirements2 and khr_dedicated_allocation

+++ common/device_memory.go
@@ -6,6 +6,6 @@ type MemoryType struct {
 }
 
 type MemoryHeap struct {
-	Size  uint64
+	Size  int
 	Flags MemoryHeapFlags
 }
